# openapi.json
{"openapi":"3.1.0","info":{"title":"pin4U API","description":"pin4U 백엔드 API 문서","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"uidCookie":[]}],"paths":{"/api/requests":{"get":{"tags":["Requests"],"summary":"요청 목록","description":"최신 순서로 요청 리스트를 반환합니다.","operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponseJava.util.MapJava.lang.StringJava.lang.Object"}}}}}},"post":{"tags":["Requests"],"summary":"요청 생성","description":"Slug가 발급되고 /r/{slug}로 접근 가능합니다.","operationId":"create","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.requests.dto.RequestDtos$CreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponseJava.util.MapJava.lang.StringJava.lang.Object"}}}}}}},"/api/requests/{slug}/recommendations":{"post":{"tags":["Recommendations"],"summary":"추천 장소 제출","description":"요청 슬러그에 대한 추천 장소들을 제출합니다.","operationId":"submit","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$SubmitRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$SubmitResponse"}}}}}}},"/api/groups":{"post":{"tags":["Groups"],"summary":"그룹 생성","description":"로그인 사용자(me)가 소유한 그룹을 생성합니다.","operationId":"create_1","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupCreateRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.groups.dto.GroupDtos$CreateResponse"}}}}},"security":[{"uidCookie":[]}]}},"/api/groups/{group_slug}/members":{"post":{"tags":["Groups"],"summary":"멤버 요청/승인","description":"`action = request | approve` (승인은 owner만 가능)","operationId":"memberAction","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"type":"string"}},{"name":"group_slug","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.groups.dto.GroupDtos$MemberActionRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.groups.dto.GroupDtos$MemberActionResponse"}}}}},"security":[{"uidCookie":[]}]}},"/api/auth/logout":{"post":{"tags":["Auth"],"summary":"로그아웃","description":"uid 쿠키를 만료시킵니다.","operationId":"logout","responses":{"200":{"description":"OK"}}}},"/api/auth/kakao/login":{"post":{"tags":["Auth"],"summary":"카카오 로그인","description":"프론트에서 전달한 Kakao access_token을 검증하고 uid 쿠키를 발급합니다.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.auth.dto.AuthDtos$KakaoLoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.auth.dto.AuthDtos$LoginResponse"}}}}}}},"/api/me":{"get":{"tags":["Member"],"summary":"내 프로필 조회","description":"uid 쿠키가 있으면 내 프로필을, 없으면 204를 반환합니다.","operationId":"me","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.member.dto.MemberDtos$UserResponse"}}}}},"security":[{"uidCookie":[]}]},"patch":{"tags":["Member"],"summary":"내 닉네임 변경","description":"2~16자 제약. 로그인(쿠키 uid) 필요.","operationId":"updateNickname","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.member.dto.MemberDtos$NicknamePatch"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.member.dto.MemberDtos$UserResponse"}}}}},"security":[{"uidCookie":[]}]}},"/api/user":{"get":{"tags":["Member"],"summary":"고정 유저 조회(데모)","description":"항상 id=1 유저를 반환합니다.","operationId":"getUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponseJava.util.MapJava.lang.StringJava.lang.Object"}}}}}}},"/api/user/1":{"get":{"tags":["Member"],"summary":"고정 유저 조회(데모)","description":"항상 id=1 유저를 반환합니다.","operationId":"getUser_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponseJava.util.MapJava.lang.StringJava.lang.Object"}}}}}}},"/api/stations/search":{"get":{"tags":["Stations"],"summary":"지하철역 검색","description":"역 이름으로 키워드 검색합니다.","operationId":"search","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.stations.dto.StationDtos$SearchResponse"}}}}}}},"/api/requests/{slug}":{"get":{"tags":["Requests"],"summary":"요청 상세","description":"지도의 핀/카드뉴스 등 상세 정보를 반환합니다.","operationId":"get","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"include_ai","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$RequestDetailResponse"}}}}}},"delete":{"tags":["Requests"],"summary":"요청 삭제","description":"요청 소유자만 삭제할 수 있습니다. 관련 집계/메모는 DB FK `ON DELETE CASCADE`로 함께 삭제됩니다.","operationId":"delete","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"type":"string"}},{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}},"security":[{"uidCookie":[]}]}},"/api/requests/{slug}/places/notes":{"get":{"tags":["Requests"],"summary":"장소 메모 조회","description":"요청 슬러그 + 외부 장소 ID로 메모를 조회합니다.","operationId":"getNotes","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}},{"name":"external_id","in":"query","required":true,"schema":{"type":"string"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponse"}}}}}}},"/api/recommendations/auto":{"get":{"tags":["Recommendations"],"summary":"자동 장소추천","description":"요청 슬러그 기준으로 자동 추천을 수행합니다.","operationId":"auto","parameters":[{"name":"slug","in":"query","required":true,"schema":{"type":"string"}},{"name":"n","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"q","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$RequestDetailResponse"}}}}}}},"/api/places/summary":{"get":{"tags":["Places"],"summary":"장소 요약(Stub)","description":"다음 단계에서 실제 DB/서비스 로직 연결 예정","operationId":"getSummary","parameters":[{"name":"external_id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.places.api.PlaceSummaryController$SummaryResponse"}}}}}}},"/api/places/search":{"get":{"tags":["Places"],"summary":"장소 검색","description":"역 코드와 쿼리로 반경 내 장소를 검색합니다.","operationId":"search_1","parameters":[{"name":"station","in":"query","required":true,"schema":{"type":"string"}},{"name":"q","in":"query","required":true,"schema":{"type":"string"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.places.dto.PlaceDtos$SearchResponse"}}}}}}},"/api/home":{"get":{"tags":["Home"],"summary":"홈 대시보드","description":"로그인 사용자의 요청 요약 목록을 반환합니다.","operationId":"get_1","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.home.dto.HomeDtos$DashboardResponse"}}}}},"security":[{"uidCookie":[]}]}},"/api/groups/{group_slug}/map":{"get":{"tags":["Groups"],"summary":"그룹지도","description":"그룹에 속한 모든 요청을 합산하여 장소 리스트를 반환합니다.","operationId":"map","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"type":"string"}},{"name":"group_slug","in":"path","required":true,"schema":{"type":"string"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.groups.dto.GroupMapDtos$Response"}}}}},"security":[{"uidCookie":[]}]}},"/api/auth/me":{"get":{"tags":["Auth"],"summary":"현재 로그인 사용자 조회","description":"uid 쿠키 기준으로 로그인 사용자를 반환합니다. 없으면 204.","operationId":"me_1","parameters":[{"name":"uid","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.auth.dto.AuthDtos$LoginUser"}}}}},"security":[{"uidCookie":[]}]}}},"components":{"schemas":{"io.github.ssforu.pin4u.features.requests.dto.RequestDtos$CreateRequest":{"type":"object","properties":{"station_code":{"type":"string"},"request_message":{"type":"string"},"group_slug":{"type":"string"}}},"io.github.ssforu.pin4u.common.response.ApiResponse$Error":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"details":{"type":"object"}}},"io.github.ssforu.pin4u.common.response.ApiResponseJava.util.MapJava.lang.StringJava.lang.Object":{"type":"object","properties":{"result":{"type":"string"},"data":{"type":"object","additionalProperties":{"type":"object"}},"error":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponse$Error"},"timestamp":{"type":"string","format":"date-time"}}},"io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$SubmitItem":{"type":"object","properties":{"external_id":{"type":"string"},"recommender_nickname":{"type":"string"},"recommend_message":{"type":"string"},"image_url":{"type":"string"},"image_is_public":{"type":"boolean"},"tags":{"type":"array","items":{"type":"string"}},"guest_id":{"type":"string"}}},"io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$SubmitRequest":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$SubmitItem"}}}},"io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$SubmitResponse":{"type":"object","properties":{"result":{"type":"string"},"data":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$SubmitResponse"},"error":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponse$Error"},"timestamp":{"type":"string","format":"date-time"}}},"io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$InvalidItem":{"type":"object","properties":{"details":{"type":"object","additionalProperties":{"type":"string"}},"external_id":{"type":"string"}}},"io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$OutOfRadiusItem":{"type":"object","properties":{"external_id":{"type":"string"},"distance_m":{"type":"integer","format":"int32"}}},"io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$SavedItem":{"type":"object","properties":{"external_id":{"type":"string"},"recommended_count":{"type":"integer","format":"int32"}}},"io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$SimpleItem":{"type":"object","properties":{"external_id":{"type":"string"}}},"io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$SubmitResponse":{"type":"object","properties":{"saved":{"type":"array","items":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$SavedItem"}},"conflicts":{"type":"array","items":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$SimpleItem"}},"out_of_radius":{"type":"array","items":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$OutOfRadiusItem"}},"not_found":{"type":"array","items":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$SimpleItem"}},"invalid":{"type":"array","items":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.recommendations.dto.RecommendationDtos$InvalidItem"}},"totals":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}}}},"GroupCreateRequest":{"type":"object","description":"그룹 생성 요청 바디 (문서용 확장 필드 포함)","properties":{"name":{"type":"string","description":"그룹 이름","example":"우리모임"},"image_url":{"type":"string","description":"그룹 이미지 URL","example":"https://example.com/group.png"},"station_name":{"type":"string","description":"(문서용) 역 이름","example":"숭실대입구"},"station_line":{"type":"string","description":"(문서용) 호선","example":"7호선"},"request_message":{"type":"string","description":"(문서용) 요청 메시지","example":"조용한 카페 찾습니다"},"group_slug":{"type":"string","description":"(문서용) 희망 그룹 슬러그(중복 불가)","example":"woori-1"}}},"io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.groups.dto.GroupDtos$CreateResponse":{"type":"object","properties":{"result":{"type":"string"},"data":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.groups.dto.GroupDtos$CreateResponse"},"error":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponse$Error"},"timestamp":{"type":"string","format":"date-time"}}},"io.github.ssforu.pin4u.features.groups.dto.GroupDtos$CreateResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"slug":{"type":"string"},"name":{"type":"string"},"image_url":{"type":"string"}}},"io.github.ssforu.pin4u.features.groups.dto.GroupDtos$MemberActionRequest":{"type":"object","properties":{"action":{"type":"string"},"user_id":{"type":"integer","format":"int64"}}},"io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.groups.dto.GroupDtos$MemberActionResponse":{"type":"object","properties":{"result":{"type":"string"},"data":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.groups.dto.GroupDtos$MemberActionResponse"},"error":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponse$Error"},"timestamp":{"type":"string","format":"date-time"}}},"io.github.ssforu.pin4u.features.groups.dto.GroupDtos$MemberActionResponse":{"type":"object","properties":{"status":{"type":"string"}}},"io.github.ssforu.pin4u.features.auth.dto.AuthDtos$KakaoLoginRequest":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]},"io.github.ssforu.pin4u.features.auth.dto.AuthDtos$LoginResponse":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.auth.dto.AuthDtos$LoginUser"},"isNew":{"type":"boolean"}}},"io.github.ssforu.pin4u.features.auth.dto.AuthDtos$LoginUser":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nickname":{"type":"string"}}},"io.github.ssforu.pin4u.features.member.dto.MemberDtos$NicknamePatch":{"type":"object","properties":{"nickname":{"type":"string"}}},"io.github.ssforu.pin4u.features.member.dto.MemberDtos$UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nickname":{"type":"string"},"preference_text":{"type":"string"},"created_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"}}},"io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.stations.dto.StationDtos$SearchResponse":{"type":"object","properties":{"result":{"type":"string"},"data":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.stations.dto.StationDtos$SearchResponse"},"error":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponse$Error"},"timestamp":{"type":"string","format":"date-time"}}},"io.github.ssforu.pin4u.features.stations.dto.StationDtos$SearchResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.stations.dto.StationDtos$StationItem"}},"count":{"type":"integer","format":"int32"}}},"io.github.ssforu.pin4u.features.stations.dto.StationDtos$StationItem":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"},"line":{"type":"string"},"lat":{"type":"number"},"lng":{"type":"number"}}},"io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$RequestDetailResponse":{"type":"object","properties":{"result":{"type":"string"},"data":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$RequestDetailResponse"},"error":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponse$Error"},"timestamp":{"type":"string","format":"date-time"}}},"io.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$Ai":{"type":"object","properties":{"summaryText":{"type":"string"},"evidence":{"type":"string"},"updatedAt":{"type":"string","format":"date-time"}}},"io.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$Item":{"type":"object","properties":{"externalId":{"type":"string"},"id":{"type":"string"},"placeName":{"type":"string"},"categoryGroupCode":{"type":"string"},"categoryGroupName":{"type":"string"},"categoryName":{"type":"string"},"addressName":{"type":"string"},"roadAddressName":{"type":"string"},"x":{"type":"string"},"y":{"type":"string"},"distanceM":{"type":"integer","format":"int32"},"placeUrl":{"type":"string"},"mock":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$Mock"},"ai":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$Ai"},"recommended_count":{"type":"integer","format":"int32"}}},"io.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$Mock":{"type":"object","properties":{"rating":{"type":"number","format":"double"},"ratingCount":{"type":"integer","format":"int32"},"imageUrls":{"type":"array","items":{"type":"string"}},"openingHours":{"type":"array","items":{"type":"string"}}}},"io.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$RequestDetailResponse":{"type":"object","properties":{"slug":{"type":"string"},"station":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$Station"},"requestMessage":{"type":"string"},"items":{"type":"array","items":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$Item"}}}},"io.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$Station":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"},"line":{"type":"string"},"lat":{"type":"number"},"lng":{"type":"number"}}},"io.github.ssforu.pin4u.common.response.ApiResponse":{"type":"object","properties":{"result":{"type":"string"},"data":{"type":"object"},"error":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponse$Error"},"timestamp":{"type":"string","format":"date-time"}}},"io.github.ssforu.pin4u.features.places.api.PlaceSummaryController$SummaryResponse":{"type":"object","properties":{"external_id":{"type":"string"},"summary":{"type":"string"},"source":{"type":"string"}}},"io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.places.dto.PlaceDtos$SearchResponse":{"type":"object","properties":{"result":{"type":"string"},"data":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.places.dto.PlaceDtos$SearchResponse"},"error":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponse$Error"},"timestamp":{"type":"string","format":"date-time"}}},"io.github.ssforu.pin4u.features.places.dto.PlaceDtos$Item":{"type":"object","properties":{"external_id":{"type":"string"},"id":{"type":"string"},"place_name":{"type":"string"},"category_group_code":{"type":"string"},"category_group_name":{"type":"string"},"category_name":{"type":"string"},"phone":{"type":"string"},"address_name":{"type":"string"},"road_address_name":{"type":"string"},"x":{"type":"string"},"y":{"type":"string"},"place_url":{"type":"string"},"distance_m":{"type":"integer","format":"int32"},"mock":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.places.dto.PlaceDtos$MockDto"}}},"io.github.ssforu.pin4u.features.places.dto.PlaceDtos$MockDto":{"type":"object","properties":{"rating":{"type":"number","format":"double"},"rating_count":{"type":"integer","format":"int32"},"review_snippets":{"type":"array","items":{"type":"string"}},"image_urls":{"type":"array","items":{"type":"string"}},"opening_hours":{"type":"array","items":{"type":"string"}}}},"io.github.ssforu.pin4u.features.places.dto.PlaceDtos$SearchResponse":{"type":"object","properties":{"station":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.places.dto.PlaceDtos$StationBrief"},"items":{"type":"array","items":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.places.dto.PlaceDtos$Item"}}}},"io.github.ssforu.pin4u.features.places.dto.PlaceDtos$StationBrief":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"},"line":{"type":"string"},"lat":{"type":"number"},"lng":{"type":"number"}}},"io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.home.dto.HomeDtos$DashboardResponse":{"type":"object","properties":{"result":{"type":"string"},"data":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.home.dto.HomeDtos$DashboardResponse"},"error":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponse$Error"},"timestamp":{"type":"string","format":"date-time"}}},"io.github.ssforu.pin4u.features.home.dto.HomeDtos$DashboardResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.requests.dto.RequestDtos$ListItem"}},"groups":{"type":"array","items":{"type":"object","additionalProperties":{"type":"object"}}},"badges":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}}}},"io.github.ssforu.pin4u.features.requests.dto.RequestDtos$ListItem":{"type":"object","properties":{"slug":{"type":"string"},"station_name":{"type":"string"},"station_line":{"type":"string"},"road_address_name":{"type":"string"},"recommend_count":{"type":"integer","format":"int32"},"created_at":{"type":"string","format":"date-time"}}},"io.github.ssforu.pin4u.common.response.ApiResponseIo.github.ssforu.pin4u.features.groups.dto.GroupMapDtos$Response":{"type":"object","properties":{"result":{"type":"string"},"data":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.groups.dto.GroupMapDtos$Response"},"error":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.common.response.ApiResponse$Error"},"timestamp":{"type":"string","format":"date-time"}}},"io.github.ssforu.pin4u.features.groups.dto.GroupMapDtos$GroupBrief":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"slug":{"type":"string"},"name":{"type":"string"},"image_url":{"type":"string"}}},"io.github.ssforu.pin4u.features.groups.dto.GroupMapDtos$Response":{"type":"object","properties":{"group":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.groups.dto.GroupMapDtos$GroupBrief"},"station_center":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.groups.dto.GroupMapDtos$StationCenter"},"items":{"type":"array","items":{"$ref":"#/components/schemas/io.github.ssforu.pin4u.features.requests.dto.RequestDetailDtos$Item"}}}},"io.github.ssforu.pin4u.features.groups.dto.GroupMapDtos$StationCenter":{"type":"object","properties":{"lat":{"type":"number"},"lng":{"type":"number"}}}},"securitySchemes":{"uidCookie":{"type":"apiKey","description":"로그인 성공 시 발급되는 uid 쿠키","name":"uid","in":"cookie"}}}}